library/requirements.txt
Django==1.8.1
MySQL-python==1.2.5
Pillow==2.8.1
django-extensions==1.5.5
djangorestframework==3.1.2
six==1.9.0
wsgiref==0.1.2
